requirements.txt
paho-mqtt==1.6.1
load-dotenv==0.1.0
mysql-connector-python==8.0.33
